+++ .github/workflows/python-package.yml
@@ -56,7 +56,6 @@ jobs:
     - name: Install dependencies
       run: |
         # sudo apt-get install gfortran swig libhdf5-serial-dev libmpich-dev
-        sudo apt-get update
         sudo apt install -y openmpi-bin libopenmpi-dev gfortran build-essential libblas3 libblas-dev liblapack3 liblapack-dev libatlas-base-dev texlive texlive-latex-extra texlive-fonts-recommended dvipng cm-super
         python -m pip install --upgrade git+https://github.com/bitranox/wrapt_timeout_decorator.git
         python -m pip install pytest pytest-cov flake8 pytest-aiohttp sqlparse freezegun PyJWT joblib tensorflow afterglowpy
