+++ nmma/em/analysis.py
@@ -16,8 +16,7 @@ from .injection import create_light_curve_data
 from .likelihood import OpticalLightCurve
 from .model import create_light_curve_model_from_args, model_parameters_dict
 from .prior import create_prior_from_args
-from .utils import getFilteredMag
-from .io import loadEvent
+from .utils import getFilteredMag, loadEvent
 
 matplotlib.use("agg")
