+++ nmma/em/create_svdmodel.py
@@ -10,8 +10,7 @@ from ..utils.models import refresh_models_list
 from . import model_parameters
 from .model import SVDLightCurveModel
 from .training import SVDTrainingModel
-from .utils import interpolate_nans
-from .io import read_photometry_files, read_spectroscopy_files
+from .utils import interpolate_nans, read_photometry_files, read_spectroscopy_files
 
 
 def axial_symmetry(training_data):
@@ -53,7 +52,7 @@ def main():
     parser.add_argument(
         "--data-file-type",
         type=str,
-        help="Type of light curve files [bulla, standard, ztf, hdf5]",
+        help="Type of light curve files [bulla, standard, ztf]",
         default="bulla",
     )
     parser.add_argument(
@@ -183,7 +182,7 @@ def main():
         )
     model_function = MODEL_FUNCTIONS[args.model]
 
-    file_extensions = ["dat", "csv", "dat.gz", "h5"]
+    file_extensions = ["dat", "csv", "dat.gz"]
     filenames = []
     for file_extension in file_extensions:
         if not args.ignore_bolometric:
