+++ nmma/em/model_parameters.py
@@ -8,15 +8,14 @@ def AnBa2022(data):
 
     data_out = {}
 
-    parameters = ["mtot", "vej", "mni", "mrp", "xmix"]
-    parameters_idx = [0, 1, 2, 3, 4]
+    parameters = ["v", "mrp", "xmix"]
+    parameters_idx = [1, 3, 4]
     magkeys = data.keys()
     for jj, key in enumerate(magkeys):
         rr = [
             np.abs(float(x))
             for x in re.findall(
-                r"[-+]?[.]?[\d]+(?:,\d\d\d)*[\.]?\d*(?:[eE][-+]?\d+)?",
-                key.replace("m56", "mni"),
+                r"[-+]?[.]?[\d]+(?:,\d\d\d)*[\.]?\d*(?:[eE][-+]?\d+)?", key
             )
         ]
 
@@ -115,6 +114,33 @@ def Bu2019lm(data):
         # Best to interpolate mass in log10 space
         rr[1] = np.log10(rr[1])
         rr[2] = np.log10(rr[2])
+
+        data_out[key] = {
+            param: rr[idx] for param, idx in zip(parameters, parameters_idx)
+        }
+        data_out[key] = {**data_out[key], **data[key]}
+
+    return data_out, parameters
+
+
+def Bu2022mv(data):
+
+    data_out = {}
+
+    parameters = ["log10_mej_dyn", "vej_dyn", "log10_mej_wind", "vej_wind", "KNtheta"]
+    parameters_idx = [0, 1, 3, 4, 6]
+    magkeys = data.keys()
+    for jj, key in enumerate(magkeys):
+        rr = [
+            np.abs(float(x))
+            for x in re.findall(
+                r"[-+]?[.]?[\d]+(?:,\d\d\d)*[\.]?\d*(?:[eE][-+]?\d+)?", key
+            )
+        ]
+
+        # Best to interpolate mass in log10 space
+        rr[0] = np.log10(rr[0])
+        rr[3] = np.log10(rr[3])
 
         data_out[key] = {
             param: rr[idx] for param, idx in zip(parameters, parameters_idx)
